move sync things to new file
